GCP-421 and GCP-427 Refactor Middleware/GermplasmBrowser/Workbench: - find functions to "get" - Return value of count methods from int or Long to long - start and numOfRows parameters from long or Integer to int - Standard unit tests output format - Standard unit tests parameter setup at the beginning of method - QueryException to MiddlewareQueryException - Standard MiddlewareQueryException message format - Addition of MiddlewareQueryExceptions to DAO methods - Addition of toString() method in some POJOs - Update of fields of some POJOs to match the table fields
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/GermplasmStudyBrowser@990 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/browser/germplasmlist/GermplasmListDataComponent.java
@@ -61,12 +61,12 @@ public class GermplasmListDataComponent extends VerticalLayout implements Initia
     @Override
     public void afterPropertiesSet() throws Exception{
         List<GermplasmListData> listData = new ArrayList<GermplasmListData>();
-        int listDataCount = this.germplasmListManager.countGermplasmListDataByListId(germplasmListId);
+        long listDataCount = this.germplasmListManager.countGermplasmListDataByListId(germplasmListId);
         if (listDataCount == 0) {
             addComponent(new Label(messageSource.getMessage(Message.no_listdata_retrieved_label))); // "No Germplasm List Data retrieved."
         } else {
             
-            listData = this.germplasmListManager.getGermplasmListDataByListId(germplasmListId, 0, listDataCount);
+            listData = this.germplasmListManager.getGermplasmListDataByListId(germplasmListId, 0, (int) listDataCount);
             
             // create the Vaadin Table to display the Germplasm List Data
             listDataTable = new Table("");
